// src/types/apps/register.ts
export type RegisterData = {
  name: string
  username: string
  email: string
  phone: string
  password: string
  password_confirmation: string
  vendor_name: string
  country_id: number
  province_id: number
  district_id: number
  subdistrict_id: number
  address: string
  business_type_id: number
}